Allow # in fragment and prohibit url from ending with '),.
Though only one # is ever allowed in a valid url per RFC 3986
the fragment is handled client side and web browsers will simply
ignore all text after the first #. Potentially text after the first #
could be relevant.

allowing # in fragments allows the merging of the path, query and fragment

Added a guard to the end of the combined path/query/fragment to prevent urls
from ending with "'),.". Though nothing in the spec prevents urls
containing those characters ' and ) are common context enders in scripts
and , . will often end a url in plain text. Fortunately the most common
context " is already dissallowed in urls.